+++ tests/test_data_catalog.py
@@ -538,6 +538,24 @@ def test_get_geodataset():
         data_catalog.get_geodataset({"name": "test"})
 
 
+def test_get_dataset(timeseries_df):
+    # get_dataset
+    data_catalog = DataCatalog("artifact_data")
+    test_dataset = timeseries_df.to_xarray()
+    subset_timeseries = timeseries_df.iloc[[0, len(timeseries_df) // 2]]
+    time_tuple = (
+        subset_timeseries.index[0].to_pydatetime(),
+        subset_timeseries.index[1].to_pydatetime(),
+    )
+    ds = data_catalog.get_dataset(test_dataset, time_tuple=time_tuple)
+    assert isinstance(ds, xr.Dataset)
+    assert ds.time[-1].values == subset_timeseries.index[1].to_datetime64()
+
+    ds = data_catalog.get_dataset(test_dataset, variables=["col1"])
+    assert isinstance(ds, xr.DataArray)
+    assert ds.name == "col1"
+
+
 def test_get_dataframe(df, tmpdir):
     data_catalog = DataCatalog("artifact_data")  # read artifacts
     n = len(data_catalog)
@@ -596,17 +614,12 @@ def test_detect_extent():
 
     # raster dataset
     name = "chirps_global"
-    bbox = (
-        11.599998474121094,
-        45.20000076293945,
-        13.000083923339844,
-        46.79985427856445,
-    )
+    bbox = 11.60, 45.20, 13.00, 46.80
     expected_temporal_range = tuple(pd.to_datetime(["2010-02-02", "2010-02-15"]))
     ds = cast(RasterDatasetAdapter, data_catalog.get_source(name))
     detected_spatial_range = to_geographic_bbox(*ds.get_bbox(detect=True))
     detected_temporal_range = ds.get_time_range(detect=True)
-    assert np.all(np.equal(detected_spatial_range, bbox))
+    assert np.allclose(detected_spatial_range, bbox)
     assert detected_temporal_range == expected_temporal_range
 
     # geodataframe