+++ docs/changelog.rst
@@ -13,11 +13,32 @@ Unreleased
 
 Added
 -----
+- Export CLI now also accepts time tuples (#660)
+
+Changed
+-------
+- Export CLI now uses '-s' for source, '-t' for time and '-i' for config. (#660)
+
+Fixed
+-----
+- Bug in `raster.set_crs` if input_crs is of type CRS. (#659)
+- Export CLI now actually parses provided geoms. (#660)
+
+Deprecated
+----------
+
+v0.9.1 (2023-11-16)
+===================
+This release contains several bugfixes to 0.9.0 as well two new CLI methods and support for STAC.
+
+Added
+-----
 - Support for exporting data catalogs to STAC catalog formats. (#617)
 - Support for reading data catalogs from STAC catalog formats. (#625)
 - Pixi is now available as an additional task runner. (#634)
 - Support exporting data from catalogs from the CLI. (#627)
 - Support for validating data catalogs from the CLI. (#632)
+- Support for validating model configurations from the CLI. (#643)
 
 
 Changed
@@ -26,9 +47,8 @@ Changed
 
 Fixed
 -----
-
-Deprecated
-----------
+- Bug in zoom level detection in `RasterDatasetAdapter` for Tifs without overviews and paths with placeholders. (#642)
+- Bug in gis_utils.meridian_offset for grids with rounding errors. (#649)
 
 v0.9.0 (2023-10-19)
 ===================